Fix security audit findings in cryptographic function wrappers
This commit addresses critical and minor issues found during a security
audit of the C function wrappers that interface with libsodium:

1. CRITICAL: Fix incorrect key size constant in auth.c
   - Changed crypto_secretbox_KEYBYTES to crypto_auth_KEYBYTES in
     pgsodium_crypto_auth_verify_by_id() function (line 112)
   - While both constants are 32 bytes, using the wrong constant is
     semantically incorrect and could break if constants change

2. CRITICAL: Fix buffer size calculation in aead.c
   - Fixed double ABYTES addition in pgsodium_crypto_aead_ietf_encrypt_by_id()
   - Changed VARSIZE_ANY() to VARSIZE_ANY_EXHDR() for correct size (line 168)
   - Removed redundant SET_VARSIZE that added ABYTES twice (line 177-178)
   - This prevents potential buffer overruns

3. Fix VARSIZE inconsistency in box.c
   - Changed VARSIZE() to VARSIZE_ANY() in pgsodium_crypto_box_seal_open()
     (line 306)
   - Ensures correct handling of short-header varlena types

4. Use size_t consistently for size variables
   - Changed int to size_t in noncegen functions in aead.c and secretbox.c
   - Improves type safety and prevents potential overflow issues

All changes maintain backward compatibility while improving security
and correctness of the cryptographic operations.

+++ src/secretbox.c
@@ -31,7 +31,7 @@ PG_FUNCTION_INFO_V1 (pgsodium_crypto_secretbox_noncegen);
 Datum
 pgsodium_crypto_secretbox_noncegen (PG_FUNCTION_ARGS)
 {
-	int         result_size = VARHDRSZ + crypto_secretbox_NONCEBYTES;
+	size_t      result_size = VARHDRSZ + crypto_secretbox_NONCEBYTES;
 	bytea      *result = _pgsodium_zalloc_bytea (result_size);
 	randombytes_buf (VARDATA (result), crypto_secretbox_NONCEBYTES);
 	PG_RETURN_BYTEA_P (result);
